requirements.txt
# Core Dependencies
playwright==1.40.0
pyyaml==6.0.1

# Progress & Logging
tqdm==4.66.1

# Utilities
python-dotenv==1.0.0

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
